Veterinario_pochitasw/Veterinario_pochitasw/backend/requirements.txt
django>=5.0
djangorestframework
drf-spectacular
psycopg2-binary
markdown
django-filter
django-cors-headers
